requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bleach-whitelist==0.0.8
bleach==2.1.3
cffi==1.11.5
click==6.7
colorama==0.3.9
cp==1.0.3
cryptography==2.2.2
django-htmlmin==0.10.0
filelock==3.0.4
Flask==1.0.2
html5lib==1.0.1
gunicorn==19.9.0
htmlmin==0.1.12
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
paramiko==2.4.1
Pillow==5.2.0
psycopg2-binary==2.7.5
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
six==1.11.0
sshtunnel==0.1.4
webencodings==0.5.1
Werkzeug==0.14.1
